[Ref/#74] refactoring formattedTime Function 시간을 string으로 바꾸는 formattedTime 함수 Helper로 이동
