+++ TurtleNeck/TurtleNeck/Presentation/Main/MainView.swift
@@ -120,8 +120,6 @@ struct MainView: View {
                     
                     let today = Calendar.current.startOfDay(for: endTime)
                     
-                    let existingStatistic = statistic.filter { $0.date == today }.first
-                    
                     if let lastStatistic = statistic.last {
                         lastStatistic.time += wearingDuration
                     }
@@ -230,20 +228,6 @@ extension MainView {
         timer?.invalidate()
         timer = nil
         timerValue = 0
-    }
-    
-    private func formattedTime(from seconds: Int) -> String {
-        let hours = seconds / 3600
-        let minutes = (seconds % 3600) / 60
-        let secs = seconds % 60
-        
-        if hours > 0 {
-            return String(format: "%02d시간 %02d분 %02d초", hours, minutes, secs)
-        } else if minutes > 0 {
-            return String(format: "%02d분 %02d초", minutes, secs)
-        } else {
-            return String(format: "%02d초", secs) 
-        }
     }
     
     private func checkBestRecord() {
