+++ TurtleNeck/TurtleNeck/Presentation/Main/RealTime/RealTimePostureView.swift
@@ -50,19 +50,3 @@ struct RealTimePostureView: View {
         Spacer()
     }
 }
-
-extension RealTimePostureView {
-    private func formattedTime(from seconds: Int) -> String {
-        let hours = seconds / 3600
-        let minutes = (seconds % 3600) / 60
-        let secs = seconds % 60
-        
-        if hours > 0 {
-            return String(format: "%02d시간 %02d분 %02d초", hours, minutes, secs) // HH:mm:ss
-        } else if minutes > 0 {
-            return String(format: "%02d분 %02d초", minutes, secs) // mm:ss
-        } else {
-            return String(format: "%02d초", secs) // ss
-        }
-    }
-}
